+++ src/modals/templates/profile.js
@@ -1,6 +1,6 @@
-import "../../components/image_picker.js";
-import spinner from "../../templates/spinner.js";
-import { __ } from '../../i18n';
+import "components/image-picker.js";
+import spinner from "templates/spinner.js";
+import { __ } from 'i18n';
 import { _converse, converse } from  "@converse/headless/core";
 import { html } from "lit-html";
 import { modal_header_close_button } from "./buttons.js";
