+++ src/plugins/chatview/bottom-panel.js
