+++ src/plugins/chatview/templates/bottom-panel.js
