+++ src/shared/chat/message-actions.js
@@ -1,6 +1,6 @@
 import log from '@converse/headless/log';
-import { CustomElement } from './element.js';
-import { __ } from '../i18n';
+import { CustomElement } from 'components/element.js';
+import { __ } from 'i18n';
 import { _converse, api, converse } from "@converse/headless/core";
 import { html } from 'lit-element';
 import { until } from 'lit-html/directives/until.js';
