+++ webpack.html
@@ -24,6 +24,7 @@
         theme: 'concord',
         show_send_button: true,
         auto_away: 300,
+        message_limit: 300,
         auto_register_muc_nickname: true,
         loglevel: 'debug',
         modtools_disable_assign: ['owner', 'moderator', 'participant', 'visitor'],
